modifs dans la métrique suivant le commentaire: The genetic algorithm was run many different times with many different fitness functions. Formulaically, the first fitness function was:
((score^3)*(frame_score)

Score is equivalent to the length of the snake minus 1 (since the snake always starts at length 1), and frame_score is the amount of frames that the snake was alive. However, originally this fitness function resulted in many snakes that looped in circles endlessly without eating any fruit to maximize the frame_score component of the function. Thus, the training was modified such that all snakes are killed off if they do not eat a fruit in 50 frames. Also, if a snake died due to not eating any fruit for 50 frames, 50 points were subtracted from the frame_score to discourage the behaviour even further.